Docker: added 1909 image

+++ docker/Update-AllBranches.ps1
@@ -67,7 +67,8 @@ foreach ($manifestName in $manifestData.Keys) {
 
     if ($PSCmdlet.ShouldProcess($manifestName, "Create manifest")) {
         # bug 954 too
-        rm -Recurse ~/.docker/manifests/*
+        if (Test-Path ~/.docker/manifests/) { rm -Recurse ~/.docker/manifests/* }
+        $env:DOCKER_CLI_EXPERIMENTAL = "enabled"
         docker manifest create $manifestName @imageList --amend
         if ($LASTEXITCODE -ne 0) {
             throw "Manifest creation error"
